chore(crons): Extract OverviewTable
